Type: fix
Description:
Click "Cancel" after adding packages on the uninstall page, no response, click "Confirm" to flash back directly.

RootCause:
After adding the package on the uninstall page, it is changed to multi-file installation, and the single-file installation page is deleted.

Solution:
Adding packages is not allowed on the uninstall page.